Fix sliding note on new click
With the previous version, the oscillator could still be ramping when you click on a new note, causing an unintended/hard-to-control sliding note. Rather than constantly ramping, only ramp when the mouse is dragged, and then immediately set the frequency on a new click. This also helps clean up some more "global" state management.