Start building drum sequencer
I'm playing around with p5, but starting to wonder if this would be easier with plain components, so I'm gonna pivot and try that, but I wanted to commit the progress I have so far since there's some stuff related to types that are interesting/potentially useful.

For example, I found an interesting example of getting a "fixed" size array type that avoids having to explicitly define a tuple type of length N (and which I would need a new type for every different value of N) - this lets me define a single type that can be used for any value of N. I put quotes around "fixed" because at runtime, it's still just an array, so there's problem some ways to break it. Still, I think it's better than using a plain array with no checks around making sure the size works as expected.