Change how beats are handled
Instead of using a Tuple (since it was proving difficult to initialize properly), just keep an array which contains the indices of beats to enable for a given track. It's not as clean, but it's a way forward so I'm not spinning my wheels so much.

Also, I extracted the DrumPad part of the sequencer into a separate component so I can add some other parts to the screen overall.