Connect oscillator playback to keyboard
Since I will need to modify envelope(s) on the Oscillator from the SoundDesign component, the Oscillator lives there and is passed into the keyboard.

Unfortunately, I had to kinda futz with the typing to get things to work so I could use an OmniOscillator so I can theoretically change Oscillator types in the future.